requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
greenlet==3.2.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest==8.3.5
python-dotenv==1.1.0
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
